refactor: split code into more reusable functions, prepare for more scrapers and update readme
