+++ config.py
@@ -3,6 +3,6 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
-APP_EMAIL = os.getenv("APP_EMAIL")
+FROM_EMAIL = os.getenv("APP_EMAIL")
 PASSWORD = os.getenv("PASSWORD")
-MY_EMAIL = os.getenv("MY_EMAIL")
+TO_EMAIL = os.getenv("MY_EMAIL")
