+++ scrapers/pracujpl_scraper.py
@@ -1,50 +1,19 @@
-from bs4 import BeautifulSoup
-import requests
-import json
 import pandas as pd
-import logging
-import sys
-
-logging.basicConfig(
-    filename="job_scraping.log",
-    level=logging.ERROR,
-    format="%(asctime)s - %(levelname)s - %(message)s",
-)
+from .utils.scraper_utils import get_search_criteria_data, get_soup
 
 
-def scrape_pracujpl():
+def scrape():
     url = _build_url()
-
-    try:
-        job_page = _fetch_job_page(url)
-    except requests.HTTPError as e:
-        logging.error(f"Failed to fetch job page from {url}: {e}")
-        sys.exit(1)
-
-    soup = BeautifulSoup(job_page, "lxml")
-
+    soup = get_soup(url)
     job_tiles = soup.find_all("div", class_="tiles_b131b74u")
-
     job_dict_list = [_extract_job_details(tile) for tile in job_tiles]
 
     return pd.DataFrame(job_dict_list)
 
 
 def _build_url():
-    with open("search_criteria.json", "r") as file:
-        data = json.load(file)
-
-    search_criteria = data["search_criteria"][0]
-    job_title = search_criteria["job_title"].lower()
-    location = search_criteria["location"].lower()
-
+    job_title, location = get_search_criteria_data()
     return f"https://it.pracuj.pl/praca/{job_title};kw/{location};wp/ostatnich%2024h;p,1?rd=0&et=17%2C1%2C3&sc=0"
-
-
-def _fetch_job_page(url: str):
-    response = requests.get(url)
-    response.raise_for_status()
-    return response.text
 
 
 def _extract_job_details(tile):
